+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current (in progress)
 
+- Fix side menu on mobile [#1701](https://github.com/opendatateam/udata/pull/1701)
 - Fix update frequency field [#1702](https://github.com/opendatateam/udata/pull/1702)
 
 ## 1.3.11 (2018-05-29)